refactor with getBookByCategory
